+++ packages/changed-elements-react/src/widgets/V2Widget/componets/styles/ChangedElementsWidget.scss
