+++ packages/changed-elements-react/src/widgets/V2Widget/models/NamedVersions.ts
@@ -1,4 +1,5 @@
-import { VersionState } from "../VersionCompareSelectWidget";
+import { VersionState } from "./VersionState";
+
 
 export interface NamedVersions {
   entries: VersionState[];
