+++ packages/changed-elements-react/src/widgets/V2Widget/models/VersionState.ts
@@ -1,0 +1,10 @@
+
+import { NamedVersion } from "../../../clients/iModelsClient";
+import { VersionProcessedState } from "../VersionProcessedState";
+import { jobStatus } from "./JobStatus";
+
+export type VersionState = {
+  version: NamedVersion;
+  state: VersionProcessedState;
+  jobStatus?: jobStatus;
+};
